Move IO code to jobReporter.py.
This change allows for less code duplication between make_job_document.py
and a compatibility version in metric-pipeline-tasks/faro.
