+++ python/lsst/verify/bin/jobReporter.py
@@ -1,5 +1,26 @@
+import json
+import time
+
 from lsst.verify import Job, MetricSet
 from lsst.daf.butler import Butler
+
+
+__all__ = ["main", "JobReporter"]
+
+
+def main(repository, collection, metrics_package, spec, dataset_name):
+    jr = JobReporter(repository,
+                     collection,
+                     metrics_package,
+                     spec,
+                     dataset_name)
+    jobs = jr.run()
+    if len(jobs) == 0:
+        raise RuntimeError('Job reporter returned no jobs.')
+    for k, v in jobs.items():
+        filename = f"{metrics_package}_{spec}_{k}_{time.time()}.json"
+        with open(filename, 'w') as fh:
+            json.dump(v.json, fh, indent=2, sort_keys=True)
 
 
 class JobReporter:
